changed indexes in linked list in lib to  size_t

+++ src/lib/lib_linked_list.h
@@ -43,7 +43,7 @@ void lib_linked_list_prepend(lib_list_t *list,lib_elem_t value);
 /// @param list the linked list that will be extended
 /// @param index the position in the list
 /// @param value the value to be inserted
-void lib_linked_list_insert(lib_list_t *list, int index,lib_elem_t value);
+void lib_linked_list_insert(lib_list_t *list, size_t index,lib_elem_t value);
 
 /// @brief Remove an element from a linked list in O(n) time.
 /// The valid values of index are [0,n-1] for a list of n elements,
@@ -52,7 +52,7 @@ void lib_linked_list_insert(lib_list_t *list, int index,lib_elem_t value);
 /// @param index the position in the list
 /// @param success a stack variable pointer to indicate success or failure
 /// @return the value removed
-lib_elem_t lib_linked_list_remove(lib_list_t *list, int index, bool *success);
+lib_elem_t lib_linked_list_remove(lib_list_t *list, size_t index, bool *success);
 
 /// @brief Retrieve an element from a linked list in O(n) time.
 /// The valid values of index are [0,n-1] for a list of n elements,
@@ -61,7 +61,7 @@ lib_elem_t lib_linked_list_remove(lib_list_t *list, int index, bool *success);
 /// @param index the position in the list
 /// @param success a stack variable pointer to indicate success or failure
 /// @return the value at the given position
-lib_elem_t lib_linked_list_get(lib_list_t *list, int index, bool *success);
+lib_elem_t lib_linked_list_get(lib_list_t *list, size_t index, bool *success);
 
 /// @brief Returns element if found in list
 /// @param list the linked list
